Ignore a lack of trailing comma in a funcdef
A trailing comma in the function definition is a syntax error if the
function definition uses the iterable unpacking operator. This syntax
error is fixed in Python 3.5, see PEP 448.

Note that the previous version of this code recognised this, but
failed to account for the * being used on parameters other than the
last, as allowed following PEP 3102.

The version is bumped to 0.1.2.