Make the variadic arguments check clearer
The variadic arguments check was improved in db158bf, it allows
functions that have keyword only arguments to be correctly linted.

The variable no_unpacking has been renamed to no_variadic_arguments to
clarify what is being checked and the references to PEP 448 and
iterable unpacking operators have been removed so as to not misslead.

The changelog has also been updated to better clarify the change.

Reviewed-by: Jakub Stasiak <jakub@stasiak.at>