Treat function calls with one argument more reasonably
In our codebase there are quite a few places where we call a function
with one, multi-line argument (a dictionary for example). Previous
flake8-strict versions would enforce the following format:

function(
    {
        'key1': 'value1',
        'key2': 'value2',
        'key3': 'value3',
    },
)

After this change the following is allowed:

function({
    'key1': 'value1',
    'key2': 'value2',
    'key3': 'value3',
})

The first form is still allowed but it's possible it'll be disallowed in
the future.