Revert "Fix trailing comma after *args/**kwargs in Python3.6+"
This reverts commit d91fc13904ad4427ac3eec2ac99d2c66e0942949.

It turns out there is still an issue in lib2to3 with the trailing comma syntax,
even though this is now support in Python 3.6+. I thought I tested this manually
but it seems I missed it. It wasn't included in the tests as it would break the
previous versions of Python.