Fix elements inside class not being linted correctly.
This was due to a change that modifies the children of a classdef node
in order to lint arguments to a class. The node is now copied instead to
avoid modifying its children.

This could also potentially affect decorators and imports, which also
have their children modified, so these have been copied in the same way.
However, as decorator nodes are separate from funcdef nodes, and
children of importfrom nodes shouldn't have to be linted in this way, it
is unlikely they would be an issue.

Fixes: #36