Use custom lib2to3 for better support
Use the blib2to3 module from black
(https://github.com/ambv/black/tree/master/blib2to3) for python versions
>= 3.6. This is a maintained fork of lib2to3 that handles trailing
commas and f-strings properly.

An alternative to this would be to wait for updates to lib2to3 (which
takes a long time), or create our own fork. I'm more inclined to use one
that is maintained and updated for now.

A grammar change means we need to include 'old_comp_for' when
determining comprehensions.

The travis config needed updating as the default setuptools didn't
support environment markers; to fix this we need to upgrade the version.